Create Exercise 1: Woof Woof
Develop a GUI using Tkinter with a class that defines the characteristics of a dog. The program should instantiate two objects from this class and assign data to its members. The program should then output the data from each object and the oldest dog should woof via a class method.