Create Exercise 4: Shapes
Develop a GUI using Tkinter to calculate the area of Shapes. Create a parent class called Shape. This should have the following methods inputSides() – Ask the user to enter the sides of the shape. Now create subclasses for a circle, rectangle, and triangle. These should include an appropriate  area() method that will use the side values from the shape class.